requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.5
django-autoslug==1.9.3
django-bootstrap3==8.2.2
django-braces==1.11.0
django-debug-toolbar==1.7
djangorestframework==3.6.2
gunicorn==19.7.1
python-decouple==3.0
pytz==2017.2
sqlparse==0.2.3
static3==0.7.0
whitenoise==3.3.0
